Address review comments: - move logic for EditorNavSidebarComponent.exitEditor out to the parent class. - turn `targetIsOpen` method into private.
